feat: Add security.md
It is generally a best practice to have a SECURITY.md document in the repo,
however this creates a maintenance burden across all of our repos. Github
allows us to have a single security policy document in our .github repo, which
is then accessible in each repo at github.com/org/repo/security/policy.

Still not everyone will be aware of this, and so adding this stub page gives
them one more way to discover the main document.